renderer/c: fix bg-cell comment + widen grapheme_utf8_ptr to usize
Two polish fixes from the SP2 Task 5 code review:
- The bg-only cell comment in encode_cells.zig had inverted logic
  ('filtered out above by isEmpty=false'); now correctly says
  isEmpty() returns false so they reach the styled branch.
- NeedsAtlasEntry.grapheme_utf8_ptr widened from u32 to usize for
  the same WASM/host-test reason FrameCtx's pointer fields are
  usize. Prevents pointer truncation in Task 6 host tests on
  64-bit.

+++ src/renderer/c/frame_ctx.zig
@@ -77,8 +77,11 @@ pub const NeedsAtlasEntry = extern struct {
     /// Offset into FrameCtx.output_buf in u32 units (i.e., index of the
     /// first u32 of this cell's 8-u32 record).
     cell_offset: u32,
-    /// Pointer into the grapheme_scratch arena.
-    grapheme_utf8_ptr: u32,
+    /// Pointer into the grapheme_scratch arena. usize (u32 on wasm32,
+    /// u64 on 64-bit host tests) for the same reason FrameCtx's
+    /// pointer fields are usize — Zig host tests need real pointers,
+    /// WASM ABI stays unchanged.
+    grapheme_utf8_ptr: usize,
     grapheme_utf8_len: u32,
     /// bit 0: bold, bit 1: italic.
     style_bits: u32,
